GraphQL/blog/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
coverage==6.5.0
decorator==5.1.1
Django==3.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.1
django-dotenv==1.4.2
django-environ==0.9.0
django-filter==2.4.0
django-graphene-permissions==0.0.4
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
exceptiongroup==1.0.4
executing==1.2.0
graphene==2.1.9
graphene-django==2.15.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
iniconfig==1.1.1
ipython==8.6.0
iso8601==1.1.0
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
text-unidecode==1.3
tomli==2.0.1
traitlets==5.5.0
wcwidth==0.2.5
